Convert abbreviations to use tpope's vim-abolish
https://github.com/tpope/vim-abolish provides some tools for more
concisely defining abbreviations. This converts from vim's built in
abbreviation command to use the Abolish command included with the
plugin.

Because the vimrc is loaded before plugins, this also required that
abbreviation definitions be moved from a file loaded at vimrc sourcing
to a file in .vim/after/plugin so that the Abolish command would be
available.